[Update] Added buggy navigation to chatroom activity, update bug fixes, & WIP paging fixes

+++ app/src/main/java/com/example/hobbyfi/ui/main/ChatroomListFragment.kt
@@ -19,18 +19,16 @@ import com.example.hobbyfi.adapters.chatroom.ChatroomListAdapter
 import com.example.hobbyfi.databinding.FragmentChatroomListBinding
 import com.example.hobbyfi.intents.ChatroomListIntent
 import com.example.hobbyfi.intents.UserIntent
-import com.example.hobbyfi.models.Chatroom
 import com.example.hobbyfi.repositories.Repository
 import com.example.hobbyfi.shared.Constants
 import com.example.hobbyfi.state.ChatroomListState
-import com.example.hobbyfi.models.User
 import com.example.hobbyfi.viewmodels.main.ChatroomListFragmentViewModel
 import com.example.spendidly.utils.VerticalSpaceItemDecoration
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.Job
+import kotlinx.coroutines.ensureActive
 import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 
 @ExperimentalCoroutinesApi
 @ExperimentalPagingApi
@@ -41,6 +39,7 @@ class ChatroomListFragment : MainFragment() {
     private lateinit var chatroomListAdapter: ChatroomListAdapter
 
     private var searchJob: Job? = null
+    private var updateJob: Job? = null
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -54,14 +53,22 @@ class ChatroomListFragment : MainFragment() {
         binding.viewModel = viewModel
 
         chatroomListAdapter = ChatroomListAdapter({ _, chatroom ->
-                // TODO: If user join chatroom is successful, delete other chatrooms from cache (+remote keys) and load only their chatroom from cache
-                // TODO: Send user update for chatroom Id and handle no connection and other errors
                 viewModel.setButtonSelectedChatroom(chatroom)
-                lifecycleScope.launch {
-                    activityViewModel.sendIntent(UserIntent.UpdateUser(mutableMapOf(
-                        Pair(Constants.CHATROOM_ID, chatroom.id.toString())
-                    )))
-                    viewModel.sendIntent(ChatroomListIntent.DeleteChatroomsCache(chatroom.id, calledFromChatroomJoin = true))
+                val userChatroomId = activityViewModel.authUser.value?.chatroomId
+
+                // if user does not have a chatroom
+                if(userChatroomId == null || userChatroomId != chatroom.id) {
+                    // If user join chatroom is successful, delete other chatrooms from cache (+remote keys) and load only their chatroom from cache
+                    lifecycleScope.launch {
+                        activityViewModel.sendIntent(UserIntent.UpdateUser(mutableMapOf(
+                            Pair(Constants.CHATROOM_ID, chatroom.id.toString())
+                        )))
+                    }
+                } else {
+                    // otherwise simply allow the user to join their chatroom
+                    updateJob = lifecycleScope.launch {
+                        joinChatroom()
+                    }
                 }
             }, {_, chatroom ->
             viewModel.setButtonSelectedChatroom(chatroom)
@@ -91,16 +98,18 @@ class ChatroomListFragment : MainFragment() {
                     binding.swiperefresh.isRefreshing = false }
             }
 
+            // initial value chatroom id check for adapter init
+            initChatroomListAdapter()
+
             activityViewModel.authUser.observe(viewLifecycleOwner, Observer { user ->
                 if(user != null) {
                     lifecycleScope.launch {
+                        Log.i("ChatroomListFragment", "user chatroom id: ${user.chatroomId}")
                         val userHasChatroom = user.chatroomId != null
 
-                        if(userHasChatroom) {
+                        if(userHasChatroom && !viewModel!!.hasDeletedCacheForSession) {
                             viewModel!!.sendIntent(ChatroomListIntent.DeleteChatroomsCache(user.chatroomId!!))
                         } else {
-                            // FIXME: Small code dup but also, like, code smell.... ew
-                            initChatroomListAdapter(userHasChatroom)
                             viewModel!!.sendIntent(ChatroomListIntent.FetchChatrooms(false))
                         }
 
@@ -118,27 +127,32 @@ class ChatroomListFragment : MainFragment() {
                                 }
                                 is ChatroomListState.ChatroomsResult -> {
                                     binding.swiperefresh.isRefreshing = false
-                                    searchJob = lifecycleScope.launch {
-                                        state.chatrooms.catch { e ->
-                                            e.printStackTrace()
-                                            if(e is Repository.ReauthenticationException) {
-                                                Toast.makeText(requireContext(), Constants.reauthError, Toast.LENGTH_LONG)
-                                                    .show()
-                                            } else  {
-                                                Log.i("ChatroomListFragment", "state.chatrooms collect() received a normal exception: $e")
-                                                Toast.makeText(requireContext(), e.message, Toast.LENGTH_LONG)
-                                                    .show()
+                                    try {
+                                        searchJob = lifecycleScope.launch {
+                                            state.chatrooms.catch { e ->
+                                                e.printStackTrace()
+                                                if(e is Repository.ReauthenticationException) {
+                                                    Toast.makeText(requireContext(), Constants.reauthError, Toast.LENGTH_LONG)
+                                                        .show()
+                                                } else  {
+                                                    Log.i("ChatroomListFragment", "state.chatrooms collect() received a normal exception: $e")
+                                                    Toast.makeText(requireContext(), e.message, Toast.LENGTH_LONG)
+                                                        .show()
+                                                }
+                                            }.collectLatest { data ->
+                                                ensureActive()
+                                                chatroomListAdapter.submitData(data)
                                             }
-                                        }.collectLatest { data ->
-                                            chatroomListAdapter.submitData(data)
                                         }
+                                    } catch(ex: Exception) {
+                                        ex.printStackTrace()
+                                        Toast.makeText(requireContext(), ex.message, Toast.LENGTH_LONG)
+                                            .show()
                                     }
                                 }
                                 is ChatroomListState.DeleteChatroomsCacheResult -> {
                                     Log.i("ChatroomListFragment", "Deleted chatrooms cache. User has a chatroom already: ${activityViewModel.authUser.value?.chatroomId}")
                                     if(!state.calledFromChatroomJoin) {
-                                        initChatroomListAdapter(userHasChatroom)
-
                                         Log.i("ChatroomListFragment", "Sending FetchChatrooms intent. User has a chatroom already: ${activityViewModel.authUser.value?.chatroomId}")
 
                                         viewModel!!.sendIntent(ChatroomListIntent.FetchChatrooms(
@@ -156,12 +170,35 @@ class ChatroomListFragment : MainFragment() {
                     }
                 }
             })
-
             return@onCreateView root
         }
     }
-    
-    private fun initChatroomListAdapter(userHasChatroom: Boolean) {
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        navController.currentBackStackEntry?.savedStateHandle?.getLiveData<Boolean>(Constants.chatroomJoined)
+            ?.observe(viewLifecycleOwner, Observer { joined ->
+                if(joined) {
+                    updateJob = lifecycleScope.launch {
+                        joinChatroom()
+                    }
+                } else {
+                    Log.i("ChatroomListFragment", "Observing user joined chatroom false")
+                }
+            })
+
+        navController.currentBackStackEntry?.savedStateHandle?.getLiveData<Boolean>(Constants.chatroomLeft)
+            ?.observe(viewLifecycleOwner, Observer { left ->
+                if(left) {
+                    leaveChatroom()
+                }  else {
+                    Log.i("ChatroomListFragment", "Observing user joined chatroom false")
+                }
+            })
+    }
+
+    private fun initChatroomListAdapter() {
         with(binding) {
             chatroomList.addItemDecoration(VerticalSpaceItemDecoration(20))
             chatroomList.adapter = chatroomListAdapter
@@ -169,19 +206,32 @@ class ChatroomListFragment : MainFragment() {
             swiperefresh.setOnRefreshListener {
                 // TODO: Paging 3 bug currently whenever endOfPagination = true in refresh() https://issuetracker.google.com/issues/174769547
                 // TODO: Attempted hack to fix it by unsubscribing from flow while the bug is fixed; HACK DOESN'T WORK AT THIS POINT
-                if(userHasChatroom || chatroomListAdapter.itemCount <= 5) {
-                    // force flow resubscription for refresh() in single auth chatroom or single REFRESH fetches reaching endOfPagination = true
-                    viewModel!!.setCurrentChatrooms(null)
-                    ChatroomListIntent.FetchChatrooms(userHasChatroom)
-                } else {
-                    chatroomListAdapter.refresh()
-                } // should trickle down to remote mediator and VM
+                searchJob?.cancel()
+                chatroomListAdapter.refresh()
+                // should trickle down to remote mediator and VM
             }
         }
     }
 
+    private suspend fun joinChatroom() {
+         viewModel.sendIntent(ChatroomListIntent.DeleteChatroomsCache(viewModel.buttonSelectedChatroom!!.id, calledFromChatroomJoin = true))
+
+        navController.navigate(
+            ChatroomListFragmentDirections.actionGlobalActivityChatroom(
+                activityViewModel.authUser.value,
+                viewModel.buttonSelectedChatroom,
+            )
+        )
+        viewModel.setButtonSelectedChatroom(null)
+    }
+
+    private fun leaveChatroom() {
+        viewModel.setButtonSelectedChatroom(null)
+    }
+
     override fun onDestroyView() {
         searchJob?.cancel() // cancel job once we need to yeet from the fragment
+        updateJob?.cancel()
         super.onDestroyView()
     }
 }
